Different limits for NSM and PA
