+++ .env.development
@@ -27,7 +27,8 @@ SIDEKIQ_WEB_UI_USERNAME=sidekiq
 SIDEKIQ_WEB_UI_PASSWORD=sidekiq
 
 INCLUDE_ACCESSIBILITY_SPECS=false
-MAX_UPLOAD_SIZE_BYTES=10485760
+NSM_MAX_UPLOAD_SIZE_BYTES=10485760
+PA_MAX_UPLOAD_SIZE_BYTES=10485760
 CLAMBY_ENABLED=true
 
 # App store credentials for testing against local app store
