+++ .env.test
@@ -17,7 +17,8 @@ SIDEKIQ_WEB_UI_PASSWORD=sidekiq
 INCLUDE_ACCESSIBILITY_SPECS=false
 
 # 5MB
-MAX_UPLOAD_SIZE_BYTES=5242880
+NSM_MAX_UPLOAD_SIZE_BYTES=5242880
+PA_MAX_UPLOAD_SIZE_BYTES=5242880
 CLAMBY_ENABLED=true
 
 APP_STORE_TENANT_ID=123
