+++ app/views/nsm/steps/supporting_evidence/edit.html.erb
@@ -16,8 +16,9 @@
           supporting_documents: @supporting_documents,
           field: :supporting_evidence,
           form: f,
+          service: :nsm,
           locale_prefix: 'nsm.steps.supporting_evidence.edit',
-          hint: t('.upload_hint', size: FileUpload::FileUploader.human_readable_max_file_size) %>
+          hint: t('.upload_hint', size: FileUpload::FileUploader.human_readable_nsm_max_file_size) %>
 
       <% if FeatureFlags.postal_evidence.enabled? %>
         <%# Postal address component %>
