+++ app/views/prior_authority/steps/alternative_quote_details/edit.html.erb
@@ -20,12 +20,12 @@
 
       <%= form.govuk_file_field :file_upload,
                                 label: { text: t(".file_upload"), size: "s" },
-                                hint: { text: t(".file_upload_hint_html", size: FileUpload::FileUploader.human_readable_max_file_size) },
+                                hint: { text: t(".file_upload_hint_html", size: FileUpload::FileUploader.human_readable_pa_max_file_size) },
                                 accept: SupportedFileTypes::SUPPORTED_FILE_TYPES.join(','),
-                                data: { max_size: ENV['MAX_UPLOAD_SIZE_BYTES'],
+                                data: { max_size: ENV['PA_MAX_UPLOAD_SIZE_BYTES'],
                                         error_heading: I18n.t('laa_multi_step_forms.errors.error_summary.heading'),
                                         size_error_message: I18n.t('shared.shared_upload_errors.attachment_too_large',
-                                                                    size: FileUpload::FileUploader.human_readable_max_file_size) } %>
+                                                                    size: FileUpload::FileUploader.human_readable_pa_max_file_size) } %>
       <% if @form_object.document_already_uploaded? %>
         <%= render "prior_authority/steps/primary_quote/uploaded_file_summary", document: @form_object.document %>
       <% end %>
