+++ app/views/prior_authority/steps/reason_why/edit.html.erb
@@ -23,10 +23,14 @@
 
       <h1 class="govuk-heading-m"><%= t(".supporting_documentation") %></h1>
       <span class="govuk-caption-m govuk-!-margin-bottom-8">
-        <%= t(".supporting_documentation_caption_html", size: FileUpload::FileUploader.human_readable_max_file_size) %>
+        <%= t(".supporting_documentation_caption_html", size: FileUpload::FileUploader.human_readable_pa_max_file_size) %>
       </span>
 
-      <%= render 'shared/multifile_upload', supporting_documents: @form_object.supporting_documents, form: form, locale_prefix: 'prior_authority.steps.reason_why.edit' %>
+      <%= render 'shared/multifile_upload',
+                 supporting_documents: @form_object.supporting_documents,
+                 form: form,
+                  service: :pa,
+                 locale_prefix: 'prior_authority.steps.reason_why.edit' %>
       <%= form.continue_button %>
     <% end %>
   </div>
