+++ helm_deploy/templates/deployment.yaml
@@ -150,8 +150,10 @@ spec:
               value: {{ .Values.variables.portalIDPMetadataFile }}
             - name: SP_ENTITY_ID
               value: {{ .Values.variables.spEntityId }}
-            - name: MAX_UPLOAD_SIZE_BYTES
-              value: {{ mul .Values.variables.maxFileUploadBytes 1 | quote }}
+            - name: NSM_MAX_UPLOAD_SIZE_BYTES
+              value: {{ mul .Values.variables.nsmMaxFileUploadBytes 1 | quote }}
+            - name: PA_MAX_UPLOAD_SIZE_BYTES
+              value: {{ mul .Values.variables.paMaxFileUploadBytes 1 | quote }}
             - name: ANALYTICS_TRACKING_ID
               valueFrom:
                 secretKeyRef:
