+++ helm_deploy/values-dev.yaml
@@ -76,7 +76,8 @@ variables:
   portalIDPMetadataFile: config/laa_portal/metadata/dev.xml
   spEntityId: LAA_Portal_CRM7_Dev
   appStore: https://main-nscc-store-dev.cloud-platform.service.justice.gov.uk
-  maxFileUploadBytes: 10485760
+  paMaxFileUploadBytes: 10485760
+  nsmMaxFileUploadBytes: 26214400
   eolUrl: https://apply-for-extension-of-upper-limits.dev.form.service.justice.gov.uk/
   enableSyncTriggerEndpoint: 'true'
   allowIndexing: 'false'
